+++ templates/device/src/my_device.cpp
@@ -137,16 +137,10 @@ int MY_QDMI_device_job_wait(MY_QDMI_Device_Job job, const size_t timeout) {
 }
 
 int MY_QDMI_device_job_get_results(MY_QDMI_Device_Job job,
-                                   QDMI_Job_Result result, const size_t size,
-                                   void *data, size_t *size_ret) {
-  return QDMI_ERROR_NOTIMPLEMENTED;
-}
-
-int MY_QDMI_device_job_get_results_for_program(MY_QDMI_Device_Job job,
-                                               const size_t program_index,
-                                               const QDMI_Job_Result result,
-                                               const size_t size, void *data,
-                                               size_t *size_ret) {
+                                   const size_t program_index,
+                                   const QDMI_Job_Result result,
+                                   const size_t size, void *data,
+                                   size_t *size_ret) {
   return QDMI_ERROR_NOTIMPLEMENTED;
 }